Switched to a different chromadb library and refactored

+++ RAG/RAG/Endpoints/DocumentCollectionModule.cs
@@ -1,4 +1,6 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using ChromaDB.Client;
+using Microsoft.AspNetCore.Mvc;
+using RAG.Common;
 using RAG.Repository;
 
 namespace RAG.Endpoints
@@ -7,20 +9,15 @@ namespace RAG.Endpoints
     {
         public static void AddDocumentCollectionEndpoints(this IEndpointRouteBuilder app)
         {
-            app.MapPost("/DocumentCollection", async (string collectionName, IEmbeddingsRepository repository) =>
+            app.MapPost("/DocumentCollection", async (string collectionName, ICollectionsRepository repository) =>
             {
+                if (string.IsNullOrEmpty(collectionName))
+                    return Results.BadRequest("Collection name cannot be empty.");
+
                 try
                 {
-                    var result = await repository.CreateDocumentCollection(collectionName);
-
-                    if (result.IsSuccess)
-                    {
-                        return Results.Ok("Collection created");
-                    }
-                    else
-                    {
-                        return Results.BadRequest(result.ErrorMessage);
-                    }
+                    await repository.CreateDocumentCollection(collectionName);
+                    return Results.Ok("Collection created");
                 }
                 catch (Exception ex)
                 {
@@ -34,20 +31,15 @@ namespace RAG.Endpoints
                 }
             });
 
-            app.MapDelete("/DocumentCollection", async (string collectionName, IEmbeddingsRepository repository) =>
+            app.MapDelete("/DocumentCollection", async (string collectionName, ICollectionsRepository repository) =>
             {
+                if (string.IsNullOrEmpty(collectionName))
+                    return Results.BadRequest("Collection name cannot be empty.");
+
                 try
                 {
-                    var result = await repository.DeleteDocumentCollection(collectionName);
-
-                    if (result.IsSuccess)
-                    {
-                        return Results.Ok("Collection deleted");
-                    }
-                    else
-                    {
-                        return Results.BadRequest(result.ErrorMessage);
-                    }
+                    await repository.DeleteDocumentCollection(collectionName);
+                    return Results.Ok("Collection deleted");
                 }
                 catch (Exception ex)
                 {
@@ -61,20 +53,14 @@ namespace RAG.Endpoints
                 }
             });
 
-            app.MapGet("/DocumentCollections", async (IEmbeddingsRepository repository) =>
+            app.MapGet("/DocumentCollections", async (ICollectionsRepository repository) =>
             {
+
+
                 try
                 {
-                    var result = await repository.ListDocumentCollections();
-
-                    if (result.IsSuccess)
-                    {
-                        return Results.Ok(result);
-                    }
-                    else
-                    {
-                        return Results.BadRequest(result.ErrorMessage);
-                    }
+                var result = await repository.ListDocumentCollections();
+                    return Results.Ok(result);
                 }
                 catch (Exception ex)
                 {
